ENH: hist_2d(): bin size increased to 32 bit

+++ trunk/host/pybar/analysis/analysis_utils.py
@@ -673,7 +673,7 @@ def hist_2d_index(x, y, shape):
     # change memory alignment for c++ library
     x = np.ascontiguousarray(x.astype(np.int32))
     y = np.ascontiguousarray(y.astype(np.int32))
-    result = np.zeros(shape=shape, dtype=np.uint16).ravel()  # ravel hist in c-style, 3D --> 1D
+    result = np.zeros(shape=shape, dtype=np.uint32).ravel()  # ravel hist in c-style, 3D --> 1D
     analysis_functions.hist_2d(x, y, shape[0], shape[1], result)
     return np.reshape(result, shape)  # rebuilt 3D hist from 1D hist
